add more tests and fix.
