+++ core/src/main/java/com/distkv/core/map/DoubleSegment.java
@@ -1,6 +1,7 @@
 package com.distkv.core.map;
 
 import com.distkv.common.utils.ByteUtil;
+import com.distkv.core.block.Block;
 
 public class DoubleSegment extends FixedValueSegment {
 
@@ -8,11 +9,16 @@ public class DoubleSegment extends FixedValueSegment {
     super(initSize, ByteUtil.SIZE_OF_DOUBLE);
   }
 
-  public double getValue(int key) {
-    return getDouble(key);
+  protected double get(int key) {
+    Block block = getBlock(key);
+    int offset = key % blockItemSize;
+    return block.readDouble(offset);
   }
 
-  public void putValue(int key, double value) {
-    put(key, value);
+  protected int put(int key, double value) {
+    Block block = getBlock(key);
+    int offset = key % blockItemSize;
+    block.write(offset, value);
+    return getAndAddPointer();
   }
 }
