+++ core/src/main/java/com/distkv/core/map/ShortSegment.java
@@ -1,6 +1,7 @@
 package com.distkv.core.map;
 
 import com.distkv.common.utils.ByteUtil;
+import com.distkv.core.block.Block;
 
 public class ShortSegment extends FixedValueSegment {
 
@@ -8,11 +9,16 @@ public class ShortSegment extends FixedValueSegment {
     super(initSize, ByteUtil.SIZE_OF_SHORT);
   }
 
-  public short getValue(int key) {
-    return getShort(key);
+  protected short get(int key) {
+    Block block = getBlock(key);
+    int offset = key % blockItemSize;
+    return block.readShort(offset);
   }
 
-  public void putValue(int key, short value) {
-    put(key, value);
+  protected int put(int key, short value) {
+    Block block = getBlock(key);
+    int offset = key % blockItemSize;
+    block.write(offset, value);
+    return getAndAddPointer();
   }
 }
